do not count ExpirationDate=time.Time{} as expired

+++ satellite/audit/verifier.go
@@ -94,7 +94,7 @@ func (verifier *Verifier) Verify(ctx context.Context, path storj.Path, skip map[
 		}
 		return Report{}, err
 	}
-	if pointer.ExpirationDate.Before(time.Now().UTC()) {
+	if pointer.ExpirationDate != (time.Time{}) && pointer.ExpirationDate.Before(time.Now().UTC()) {
 		errDelete := verifier.metainfo.Delete(ctx, path)
 		if errDelete != nil {
 			return Report{}, Error.Wrap(errDelete)
@@ -368,7 +368,7 @@ func (verifier *Verifier) Reverify(ctx context.Context, path storj.Path) (report
 		}
 		return Report{}, err
 	}
-	if pointer.ExpirationDate.Before(time.Now().UTC()) {
+	if pointer.ExpirationDate != (time.Time{}) && pointer.ExpirationDate.Before(time.Now().UTC()) {
 		errDelete := verifier.metainfo.Delete(ctx, path)
 		if errDelete != nil {
 			return Report{}, Error.Wrap(errDelete)
@@ -436,7 +436,7 @@ func (verifier *Verifier) Reverify(ctx context.Context, path storj.Path) (report
 				verifier.log.Debug("Reverify: error getting pending pointer from metainfo", zap.Binary("Segment", []byte(pending.Path)), zap.Stringer("Node ID", pending.NodeID), zap.Error(err))
 				return
 			}
-			if pendingPointer.ExpirationDate.Before(time.Now().UTC()) {
+			if pendingPointer.ExpirationDate != (time.Time{}) && pendingPointer.ExpirationDate.Before(time.Now().UTC()) {
 				errDelete := verifier.metainfo.Delete(ctx, pending.Path)
 				if errDelete != nil {
 					verifier.log.Debug("Reverify: error deleting expired segment", zap.Binary("Segment", []byte(pending.Path)), zap.Stringer("Node ID", pending.NodeID), zap.Error(errDelete))
